Add `OptionalObjectReader`
This is essentially identical to `AnyObjectReader`, except that it
allows the top level object to be null as well.

The documentation for `AnyObjectReader.Read()` now makes it clear that
while elements of nested lists and dictionaries are allowed to be
`null`, the top level item is not.

Both `AnyObjectReader` and `OptionalObjectReader` now also expose their
functionality as a static `Read()` method, which takes two boolean flags
(`allowNullInArrays` and `allowNullAsPropertyValue`, both defaulting to
`true`), giving more control over where `null` is allowed. The instance
method just calls through to this new static one to retain the current
behaviour.

The version number is bumped to 7.1.0 due to the new API.